+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- changed: (Hedera) Save memo to queried transactions
+- fixed: (Hedera) Fix missing transactions by specifying query limit
+
 ## 4.48.1 (2025-05-27)
 
 - fixed: Do not send duplicate new-transaction notifications.